rename(Package): from ppa to src to match CI configuration
